+++ lib/dropdown_preference.dart
@@ -52,7 +52,6 @@ class _DropdownPreferenceState extends State<DropdownPreference> {
   }
 
   onChange(var val) {
-    if (widget.onChange != null) widget.onChange(val);
     if (val is String) {
       this.setState(() => PrefService.setString(widget.localKey, val));
     } else if (val is int) {
@@ -62,5 +61,7 @@ class _DropdownPreferenceState extends State<DropdownPreference> {
     } else if (val is bool) {
       this.setState(() => PrefService.setBool(widget.localKey, val));
     }
+    PrefService.notify(widget.localKey);
+    if (widget.onChange != null) widget.onChange(val);
   }
 }